Created the models table in create_tables.sql
Includes model_id for primary key.

Includes name set to NVARCHAR to allow for any number of symbols and languages in naming conventions via importing. Set to NOT NULL to require a name.

Includes source_url, which is also set to NOT NULL due to the use-case being to grab existing models via API, meaning that all models should have some kind of url pointing back to them.

Included license_type, set to NVARCHAR to account for different languages or symbols

Lastly, included description, which is set to NVARCHAR(max) to allow for most lengths and all languages/symbols used in Thingiverse descriptions.