Created a printlog create table.
Includes:

print_id = Primary Key
model_id = foreign key reference for specific models
material_used = NVARCHER(100) to account for all languages and symbols in possible material, and is set to NOT NULL since all printed models will be printed with some kind of material.

print_status = NVARCHAR(50) to log whether the print was successful, failed, partial, or cancelled.

print date = DATE, and is not null due to there always being a print date.

Duration minutes = Int, not NULL, due to all print projects having some kind of duration time. If the duration time is unknown, -1.

Lastly, already mentioned, but there's Foreign Key for model_id, so it connects to existing models in the database that are being printed via the log.