Created a settings.json file to make sure all tabs are set to 4 spaces, and that spaces are used.
Added the create or alter procedure to ThreeDModelTrackerDB, as well as the validation for the NOT NULL parameters.

Added validation that trims leading and trailing spaces, and checks if the model name is empty. If so, it raises an error and returns -1.

For Source URL, if that is empty, raises an error, and it returns -2.