I was going to move on to assign_tags.sql, but then realized I made a silly mistake with add_tag.sql and forgot to add some pretty important behavior. I started to add to it, with adding an error handle check to make sure that tag name is not null. It mirrors what we did with add_model.
However, things are starting to get out of hand with inconsistencies, including error codes and what each file is supposed to return on success, etc. I need to take a minute and draw out a quick plan and make sure all of my procedures follow the same template before I move on to the next thing.

I also need to make sure I'm accounting for any other errors that could occur, like simultaneous adds creating race conditions and the like.

I'm not too worried about this, I just need to sit down when I have a minute and do it.