Made a few final changes to add_tag,sql. I missed changing AddTag in the CREATE OR ALTER PROCEDURE line to dbo.AddTag.
I also removed the Raise Error that I have to avoid using exceptions for flow control. That dialogue can be explained from the program side of things.  Now I'm simply Returning 1, for a Validation fail, and setting the TagID to NULL for predictable behavior.

Lastly, renamed the file from the old snake_case to more conventional PascalCase for file names, although I've seen some differing opinions on that.. I've seen some use snake_case and give those directions, and others suggest camelCase or PascalCase. For now, I'm going with this as it seems to be the convention for T-SQL in many circumstances.

We're ready to move on to refactoring add model!