Made a few changes to CreateTables
Added a Unique identifier to SourceURL so it will protect against race conditions, etc. However, a NVARCHAR(MAX) can't be unique, so I set it to a more reasonable 450. This should be safe for urls. If we somehow run into an issue, we'll know why.

I also added ON DELETE CASCADE to the foreign keys in a couple tables, to get rid of any possible hanging information from past items that are deleted from the tables.

Lastly, I added a RequestID, which is a not nullable unique key. I debated with myself if it should be NULL or not, but I do feel like NOT NULL makes the most sense. I can't think of a current reason why I would ever have a system that can't simply create a request ID for a print log, but if I run into it later, or if I create some trusted system that I for some reason need to make this nullable for, I can come back and change it before the final round.