Updated RecordModelPrint to match changes we made in Create Tables, mainly adding a RequestID parameter as UNIQUEIDENTIFIER for guid, to prevent race conditions/duplicate entries from multiple button presses and the like.
However, we still need a condition for trying to add the same entry multiple times manually, either accidentally or maliciously. Currently, RecordID won't do this on its own. It will only prevent against duplicate button presses, but if the logs are made separetely and entered separately, then a new guid will be created for each entry, making them two different entries. We need a way to check for this.

I will consider the options and likely combine a bunch of stuff together and create that now.