Created a test script for GetTags. Follows same pattern as GetModels, GetPrinters, etc. Ran the script, and it's working as intended.
Also did some light work with my testing files. Added a very basic header to the top of every file, and added a use for my testing database, so it always uses that database when I run the tests, just in case.

Added the tester as completed in my roadmap.