Created an UpdateModel file that will allow the user to update models by including a model id and updating certain aspects of it, like license type, model description, and url.
Checks validation for parameters being null, checks that the model actually exists in the system, and checks if the sourceURL being included to update is already being in use by a DIFFERENT model (not equal to modelID of current model).

Also updated the roadmap with the past couple of changes.