Created a test file for UpdateModel that checks if the procedure is working correctly.
It wipes the database tables of all previous models, and then creates test models.

Then, it selects and shows the initial state of the models.

Then it tests that it can successfully update one of the models correctly, and shows that state.

It also provides validation errors for the model not existing, and trying to change the calibration cube (one of the model's) url to the existing benchy url. It also provides a validation error for having a null name.

Tested, and is working as intended at the moment. Will continue forward, and clean up the tests in the near future.