Created the last update procedure, UpdatePrintLog!
It checks for these predictable outcomes:

Null printID, or any other required parameter (ModelID, PrinterID, StartTime, EndTime, MaterialUsed, PrintStatus), and returns 1.

Checks if the PrintID to update actually exists. If not, returns 2.

Checks if the foreign keys for ModelID or PrinterID actually exist. If not, it returns 3 or 4, respectably.

Lastly, if the ModelID/StartDateTime combo is already in use by a different PrintLog, return 5.

Otherwise, return 0 for success, or Throw for an unexpected outcome!

Updated the roadmap to reflect the addition.