Created a tester for the DeleteModel procedure.
These deletes are fairly straight forward. Deletes the data from the tables, then creates test data for ModelID, TagId, and PrinterID, in order to test that even when tags are assigned to a model, or a model is used in a PrintLog, that the details will be deleted as well.

Test 1: Successfully updates and deletes the Model. Should return 0.

Test 2: ModelID does not exist, so return 2.

Test 3: ModelID is null, so return 1.

Final verification proves that the PrintLog and tag tables are also wiped when the model is deleted, due to ON CASCADE. So far, works as intended!