Created a DeleteTag and its tester, TestDeleteTag.
The procedure is pretty standard. Just checks if the input is null (return 1), or if the TagID actually exists (return 2 if not). Otherwise, it returns 0 for success, or throws for unexpected error.

Tester deletes table data, creates Model and Tag test data, and then tests for validation.

test 1: Successful delete, should return 0.

test 2: TagID does not exist, so should return 2.

Null TagId, should return 1.

Everything is working as expected!