Created a DeletePrinter procedure and its tester.
The procedure, like Delete Model, is pretty straightforward.

If the Printer is Null, return 1.

If the printer does not actually exist, Return 2.

Otherwise, delete the printer and return 0, or throw an unexpected failure.

The tester tests these potential outcomes:

First, it wipes the table and creates new tester data for PrinterID, ModelId, and creates a PrintLog using both model ids.

Test 1: Successfully deleted! Should return 0, and remove the printer.

Test 2: PrinterId not found: Should return 2.

Test 3: Validation Failure, due to printer being NULL. Should return 1.

Lastly, it shows tables to validate the removal, including that the PrintLog was successfully removed due to ON CASCADE.

Everything is working as intended!