Created DeletePrintLog, along with its tester, TestDeletePrintLog.
The delete procedure, as usual, is pretty straightforward.

If printID is null, return 1.

If PrintID does not exist, the Record doesn't exist, so return 2.

Otherwise, return 0 on successful deletion, or throw for unexpected failures.

For the tester, deletes data from the tables, and creates tester data for Model and Printer, and makes records between them.

Test 1: Tests that the record was successfully deleted. Should return 0.

Test 2: Tests that the PrintID does not exist, so returns 2.

Test 3: Tests that the PrintID is null, so should return 1.

Views all print logs so you can see the changes, including the print log that is successfully deleted.

Everything is working as intended!