Created a delete procedure (update from the user's perspective) for RemoveTagFromModel, and also its tester, TestRemoveTagFromModel.
I also updated the roadmap to reflect that, yes, the final CRUD operations seem to be completed! Yay!

Anyway, onto the procedure:

Simply removes a tag from a model. Takes a ModelID and a TagID. If the ModelId or Tag ID is NULL, return 1.

If the link between Tag and Model does not exist, return 2.

Otherwise, return 0 on success, or throw for unexpected error.

The tester:

First, deletes old data from tables, and then creates new tester data.

Then it creates links between Model and Tag using AssignTagToModel.

After this is done, it tests:

Test 1: Successfully remove a tag from a model, and end their relationship. Should return 0.

Test 2: The link between the tag and model does not exist, so should return 2.

Test 3: ModelID is NULL, so return 1.

Test 4: TagID is null, so return 1.

At the end, views all tables so we can see the results.

So far, everything is working as intended!