Well, that's what I get for committing first, checking later lol. Turns out there's some conflicts with using complex case statements with OFFSET/FETCH, and I couldn't get it to work in my server. With that in mind, I went deep diving into Dynamic querying and worked with that to refactor my GetModels file operations completely.
First, it declares variables to house the important bits for dynamic querying: a variable for the main query string, and a variable for sort by parameters that also helps prevent SQL injection by no-good villains.

It then sets OrderByString by using the SortBy optional parameter to determine if it should be sorted by ModelID, ModelName, or LicenseType. If none of them match, then it defaults to ModelName.

Then it also sets another case statement to determine sort direction based on the optional parameter SortDirection. It stores both of these in OrderByStrin.

Next, it sets the SQL variable to a dynamic query string that includes OrderByString with an expression.

At the end, it uses sp_executesql to dynamically set these sorting requirements ahead of time, allowing OFFSET and FETCH to do its thing.

It will take me a bit of time to wrap my head around a tester for this, but it will hopefully be pretty straightforward. Just need to create a LOT of testing elements to make sure things like pagination and sorting are working correctly.