Refactored TestGetPrintLogsForPrinter to test for optional parameters for filtering, pagination, and sorting.
Adds a test to check for printers with no print log history (should be an empty table).

Have tested in SSMS and everything is working as intended.

We are so close to finishing refactoring all of our get methods! Definitely powering through to get to the end of this section!

updated the roadmap as well.