Refactored both GetModelDetails and TestGetModelDetails, to include optional parameters for sorting, filtering, and pagination.
Because Get Model Details returns 3 separate tables, there needed to be separate pagination for all the tables. The first table is just one row, so there's nothing to sort or worry about, but I added pagination, sorting, and filtering for the tags assigned to the model and the print logs associated with the model.

For the tester, I refactored it to test that functionality. I tried playing around with while loops to see how that would work with creating test content, but I ran into some complex issues in SSMS. Regardless, I realized that using while loops wouldn't really work that well, because the readability would suck for the testing data (tags would essentially be Tag 1, Tag 2, Tag 3" and models would be "Model 1, etc." I would rather have very readable test data, even if it technically violates dry by having tons of "AddTags" and such.

Regardless, ran this in SSMS, and it worked!

We've now completed every single Get procedure! I am quite happy, and am now going to take a break to refresh my brain since I'm wiped.

Updated the roadmap to reflect the changes.