Added a view procedure for Recent3DPrints. It's pretty simple, as it just takes the last 10 recent model prints and shows the information. Model Name, Printer Brand , Printer Model Name, Print Status.
Also created the tester for it, and tested it in SSMS. It is working well! Wipes the tables, creates test data for a couple models and a printer, creates print logs for 12 prints, and then shows that the view only shows the most recent 10.

Tested in SSMS, and it's working!