Created a view for Model Print Summary. Shows a list of models and model ids, how many times they've been printed, the success and failure rates of those printings, and the average duration of minutes of those prints. Matches PrinterPerformanceSummary.
I also created a tester for it. As usual, it wipes the tables clean, and then creates some tester models and logs before showing the view.

Tested in SSMS and working perfectly!

Updated the roadmap to reflect the work completed.

Now I need to go to work, so I'll have to leave the last view and the rest of this for later, but I'm happy with what I was able to get done this morning.