Found a missing @ in GetModels, which was creating a bug where it would only return empty tables. Fixed the bug on line 73, where LicenseFilter is now @LicenseFilter.
I found this bug because I also added a couple tests to the TestGetModels script. The two tests are:

test 5: Filtering models with a GPL license. This proves that not only will it filter based on GPL behind included, but it will also find them even if they are different, like GPLv3 and GPL version 3.

The last test does something similar with MIT, except it also tests sorting and pagination as well. It returns results from page 2, and sorts by name, while also returning any models iwth license type MIT.

Everything was tested in SSMS and is working correctly.