Worked on my create tables script to perform a couple of new, professional functions to make it more idempotent:
First, it checks if ThreeDModelsTrackerDB exists in someone's server. If it does, continue. Otherwise, create the database.

After that, it switches to that database, to avoid creating tables in a master database.

Lastly, I added a check for each table, to see if the table already exists in the database. If it does, continue on. If it doesn't, create it.

This will avoid server errors from occuring, and allow someone to run it multiple times, either to create additional tables, or just a sanity check to make sure all tables are what they should be.

Ran in SSMS and everything seems to be correct.