Deleted the setup.sql file I created and made a build.bat file. It just was way too buggy when using SSMS. I think it's a known bug/finnickiness with SSMS and file paths or something. Either way, I think it had to do something with me using a variable for my absolute paths, and hardcoding it wouldn't make it mobile or reusable, so I just said screw it and went in a different direction.
The build.bat file will auto-create a Deploy.sql file that will then be run in SSMS. The deploy file will be a monolithic file that includes all procedures that need to be run.

I tested it, and it works well. I need to update my setup instructions and such, and also need to think about my scripts in general. I have a delete tables and delete database script in there, and I'm not comfortable having them be added to the project database so flippantly. I'll look at my options.